requirements.txt
torch
clip
faiss-cpu
pandas
pillow
flask
flask-cors
requests
openpyxl